refactor(core): consolidate core-* modules by domain boundary
Merge 12 single-purpose modules into their domain owners across core-process,
core-infra, core-unified-agent, core-agent and core-ai-gateway. Package subpath
exports are repointed so consumer import specifiers stay unchanged; Fleet Console
source aliases follow the repointed subpaths.

Relocate the executor port runtime singletons next to the executor session
manager and record why they are the fleet-admiral registration channel.
