--- packages/core-infra/src/data-dir/index.ts
@@ -1,3 +1,0 @@
-// data-dir 공개 배럴 — named export 일원화 (집계 객체 없음)
-
-export { getFleetDataDir } from "./paths.js";
