--- packages/core-infra/src/data-dir/settings/index.ts
@@ -1,10 +1,0 @@
-// global-options 공개 배럴 — 내부 구현용 심볼(createEmptyGlobalOptionsData)은 비공개
-
-export { createGlobalOptionsService } from "./service.js";
-export { createGlobalOptionsStore, sanitizeGlobalOptionsData } from "./store.js";
-export type {
-  GlobalOptionsData,
-  GlobalOptionsService,
-  GlobalOptionsStore,
-  GlobalOptionsValidationResult,
-} from "./types.js";
