+++ packages/core-infra/src/index.ts
@@ -1,5 +1,5 @@
 import { createAuthService, DEFAULT_AUTH_PATH } from "./auth/auth-storage.js";
-import { createGlobalOptionsService, type GlobalOptionsService } from "./data-dir/settings/index.js";
+import { createGlobalOptionsService, type GlobalOptionsService } from "./data-dir/settings/store.js";
 import type { AuthService } from "./auth/types.js";
 
 export interface InfraServices {
@@ -12,10 +12,20 @@ export interface InfraServicesDeps {
 }
 
 export * from "./auth/index.js";
-export * from "./data-dir/index.js";
+export * from "./data-dir/paths.js";
 export * from "./fs-store/index.js";
-export * from "./workspace-dir/index.js";
-export * from "./data-dir/settings/index.js";
+export * from "./workspace-dir/workspace-dir.js";
+export {
+  createGlobalOptionsService,
+  createGlobalOptionsStore,
+  sanitizeGlobalOptionsData,
+} from "./data-dir/settings/store.js";
+export type {
+  GlobalOptionsData,
+  GlobalOptionsService,
+  GlobalOptionsStore,
+  GlobalOptionsValidationResult,
+} from "./data-dir/settings/store.js";
 
 export function createInfraServices(deps: InfraServicesDeps = {}): InfraServices {
   const globalOptionsService = createGlobalOptionsService();
