--- packages/core-infra/src/workspace-dir/index.ts
@@ -1,12 +1,0 @@
-export {
-  ensureWorkspaceDirectory,
-  findWorkspaceDirectory,
-  getWorkspaceDirectoryRoot,
-  resolveWorkspaceDirectory,
-  resolveWorkspaceDirectoryByName,
-  toWorkspaceDirectoryName,
-} from "./workspace-dir.js";
-export type {
-  WorkspaceDirectory,
-  WorkspaceDirectoryIdentity,
-} from "./types.js";
