--- packages/core-process/src/hidden.ts
@@ -1,3 +1,0 @@
-export function withHidden<T extends object>(options?: T): T & { windowsHide: true } {
-  return { ...(options ?? {}), windowsHide: true } as T & { windowsHide: true };
-}
