+++ packages/core-process/src/index.ts
@@ -1,3 +1,2 @@
 export type { ResolvedBinary, ResolveBinaryOptions } from "./bin-resolver.js";
-export { createChildEnv, findBinaryPath, prependPathEntries, resolveBinary, resolvePathBinary } from "./bin-resolver.js";
-export { withHidden } from "./hidden.js";
+export { createChildEnv, findBinaryPath, prependPathEntries, resolveBinary, resolvePathBinary, withHidden } from "./bin-resolver.js";
