+++ packages/core-unified-agent/src/client/UnifiedCursorAgentClient.ts
@@ -21,7 +21,7 @@ import type {
   AcpToolCall,
   AcpToolCallUpdate,
 } from '../types/acp.js';
-import type { ConnectionState, StructuredLogEntry } from '../types/common.js';
+import type { ConnectionState, StructuredLogEntry } from '../types/config.js';
 import type {
   ConnectResult,
   ConnectionInfo,
@@ -37,8 +37,8 @@ import {
   mcpServerConfigsToAcp,
 } from '../config/CliConfigs.js';
 import { cleanEnvironment } from '../utils/env.js';
-import { getCursorSpawnEffortInfo, getProviderModels } from '../models/ModelRegistry.js';
-import type { ProviderModelInfo } from '../models/schemas.js';
+import { getCursorSpawnEffortInfo, getProviderModels } from '../models/registry.js';
+import type { ProviderModelInfo } from '../models/registry.js';
 
 interface StoredCursorConnectOptions extends UnifiedClientOptions {
   cli: 'cursor';
