--- packages/core-unified-agent/src/service-status/index.ts
@@ -1,5 +1,0 @@
-export type {
-  ServiceSnapshot,
-  HealthStatus,
-  ProviderKey,
-} from './types.js';
